+++ src/components/index.js
@@ -1,0 +1,8 @@
+import React from "react";
+import Piechart from "./piechart";
+
+const AllCharts = () => {
+    return <Piechart />
+}
+
+export default AllCharts
